requirements.txt
streamlit
opencv-python
mediapipe
numpy
pandas
pillow
reportlab